village js refactor
 Changes to be committed:
	modified:   ui_components/build/village/KageQuarters.js
	modified:   ui_components/build/village/Village.js
	modified:   ui_components/build/village/VillageHQ.js
	modified:   ui_components/build/village/WarTable.js
	modified:   ui_components/build/village/WorldInfo.js
	modified:   ui_components/src/village/KageQuarters.js
	modified:   ui_components/src/village/Village.js
	modified:   ui_components/src/village/VillageHQ.js
	modified:   ui_components/src/village/WarTable.js
	modified:   ui_components/src/village/WorldInfo.js
