+++ ui_components/src/village/KageQuarters.js
@@ -1,7 +1,5 @@
 import { apiFetch } from "../utils/network.js";
 import { StrategicInfoItem } from "./StrategicInfoItem.js";
-import { useModal } from '../utils/modalContext.js';
-import { TradeDisplay } from "./TradeDisplay.js";
 
 export function KageQuarters({
     playerID,
@@ -46,238 +44,280 @@ export function KageQuarters({
         { resource_id: 3, resource_name: "wealth", count: 0 }
     ]);
     const [requestedRegions, setRequestedRegions] = React.useState([]);
-    const { openModal } = useModal();
+    const [modalState, setModalState] = React.useState("closed");
+    const [modalHeader, setModalHeader] = React.useState(null);
+    const [modalText, setModalText] = React.useState(null);
     const ChangePolicy = () => {
-        apiFetch(
-            villageAPI,
-            {
-                request: 'CreateProposal',
-                type: 'policy',
-                policy_id: displayPolicyID,
-            }
-        ).then((response) => {
-            if (response.errors.length) {
-                handleErrors(response.errors);
-                return;
-            }
-            setProposalDataState(response.data.proposalData);
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
+        if (modalState == "confirm_policy") {
+            apiFetch(
+                villageAPI,
+                {
+                    request: 'CreateProposal',
+                    type: 'policy',
+                    policy_id: displayPolicyID,
+                }
+            ).then((response) => {
+                if (response.errors.length) {
+                    handleErrors(response.errors);
+                    return;
+                }
+                setProposalDataState(response.data.proposalData);
+                setModalHeader("Confirmation");
+                setModalText(response.data.response_message);
+                setModalState("response_message");
             });
-        });
+        }
+        else {
+            setModalState("confirm_policy");
+            setModalHeader("Confirmation");
+            setModalText("Are you sure you want to change policies? You will be unable to select a new policy for 3 days.");
+        }
     }
     const DeclareWar = () => {
-        apiFetch(
-            villageAPI,
-            {
-                request: 'CreateProposal',
-                type: 'declare_war',
-                target_village_id: strategicDisplayRight.village.village_id,
-            }
-        ).then((response) => {
-            if (response.errors.length) {
-                handleErrors(response.errors);
-                return;
-            }
-            setProposalDataState(response.data.proposalData);
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
+        if (modalState == "confirm_declare_war") {
+            apiFetch(
+                villageAPI,
+                {
+                    request: 'CreateProposal',
+                    type: 'declare_war',
+                    target_village_id: strategicDisplayRight.village.village_id,
+                }
+            ).then((response) => {
+                if (response.errors.length) {
+                    handleErrors(response.errors);
+                    return;
+                }
+                setProposalDataState(response.data.proposalData);
+                setModalText(response.data.response_message);
+                setModalHeader("Confirmation");
+                setModalState("response_message");
             });
-        });
+        }
+        else {
+            setModalState("confirm_declare_war");
+            setModalText("Are you sure you declare war with " + strategicDisplayRight.village.name + "?");
+            setModalHeader("Confirmation");
+        }
     }
     const OfferPeace = () => {
-        apiFetch(
-            villageAPI,
-            {
-                request: 'CreateProposal',
-                type: 'offer_peace',
-                target_village_id: strategicDisplayRight.village.village_id,
-            }
-        ).then((response) => {
-            if (response.errors.length) {
-                handleErrors(response.errors);
-                return;
-            }
-            setProposalDataState(response.data.proposalData);
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
+        if (modalState == "confirm_offer_peace") {
+            apiFetch(
+                villageAPI,
+                {
+                    request: 'CreateProposal',
+                    type: 'offer_peace',
+                    target_village_id: strategicDisplayRight.village.village_id,
+                }
+            ).then((response) => {
+                if (response.errors.length) {
+                    handleErrors(response.errors);
+                    return;
+                }
+                setProposalDataState(response.data.proposalData);
+                setModalText(response.data.response_message);
+                setModalHeader("Confirmation");
+                setModalState("response_message");
             });
-        });
+        }
+        else {
+            setModalState("confirm_offer_peace");
+            setModalText("Are you sure you want to offer peace with " + strategicDisplayRight.village.name + "?");
+            setModalHeader("Confirmation");
+        }
     }
     const OfferAlliance = () => {
-        apiFetch(
-            villageAPI,
-            {
-                request: 'CreateProposal',
-                type: 'offer_alliance',
-                target_village_id: strategicDisplayRight.village.village_id,
-            }
-        ).then((response) => {
-            if (response.errors.length) {
-                handleErrors(response.errors);
-                return;
-            }
-            setProposalDataState(response.data.proposalData);
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
+        if (modalState == "confirm_form_alliance") {
+            apiFetch(
+                villageAPI,
+                {
+                    request: 'CreateProposal',
+                    type: 'offer_alliance',
+                    target_village_id: strategicDisplayRight.village.village_id,
+                }
+            ).then((response) => {
+                if (response.errors.length) {
+                    handleErrors(response.errors);
+                    return;
+                }
+                setProposalDataState(response.data.proposalData);
+                setModalText(response.data.response_message);
+                setModalHeader("Confirmation");
+                setModalState("response_message");
             });
-        });
+        }
+        else {
+            setModalState("confirm_form_alliance");
+            setModalText("Are you sure you want to form an alliance with " + strategicDisplayRight.village.name + "?\nYou can be a member of only one Alliance at any given time.");
+            setModalHeader("Confirmation");
+        }
     }
     const BreakAlliance = () => {
-        apiFetch(
-            villageAPI,
-            {
-                request: 'CreateProposal',
-                type: 'break_alliance',
-                target_village_id: strategicDisplayRight.village.village_id,
-            }
-        ).then((response) => {
-            if (response.errors.length) {
-                handleErrors(response.errors);
-                return;
-            }
-            setProposalDataState(response.data.proposalData);
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
+        if (modalState == "confirm_break_alliance") {
+            apiFetch(
+                villageAPI,
+                {
+                    request: 'CreateProposal',
+                    type: 'break_alliance',
+                    target_village_id: strategicDisplayRight.village.village_id,
+                }
+            ).then((response) => {
+                if (response.errors.length) {
+                    handleErrors(response.errors);
+                    return;
+                }
+                setProposalDataState(response.data.proposalData);
+                setModalText(response.data.response_message);
+                setModalHeader("Confirmation");
+                setModalState("response_message");
             });
-        });
+        }
+        else {
+            setModalState("confirm_break_alliance");
+            setModalText("Are you sure you want break an alliance with " + strategicDisplayRight.village.name + "?");
+            setModalHeader("Confirmation");
+        }
     }
     const CancelProposal = () => {
-        apiFetch(
-            villageAPI,
-            {
-                request: 'CancelProposal',
-                proposal_id: currentProposal.proposal_id,
-            }
-        ).then((response) => {
-            if (response.errors.length) {
-                handleErrors(response.errors);
-                return;
-            }
-            setProposalDataState(response.data.proposalData);
-            setCurrentProposal(null);
-            setCurrentProposalKey(null);
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
+        if (modalState == "confirm_cancel_proposal") {
+            apiFetch(
+                villageAPI,
+                {
+                    request: 'CancelProposal',
+                    proposal_id: currentProposal.proposal_id,
+                }
+            ).then((response) => {
+                if (response.errors.length) {
+                    handleErrors(response.errors);
+                    return;
+                }
+                setProposalDataState(response.data.proposalData);
+                setCurrentProposal(null);
+                setCurrentProposalKey(null);
+                setModalText(response.data.response_message);
+                setModalHeader("Confirmation");
+                setModalState("response_message");
             });
-        });
+        }
+        else {
+            setModalState("confirm_cancel_proposal");
+            setModalText("Are you sure you want to cancel this proposal?");
+            setModalHeader("Confirmation");
+        }
     }
     const OfferTrade = () => {
-        apiFetch(
-            villageAPI,
-            {
-                request: 'CreateProposal',
-                type: 'offer_trade',
-                target_village_id: strategicDisplayRight.village.village_id,
-                offered_resources: offeredResources,
-                offered_regions: offeredRegions,
-                requested_resources: requestedResources,
-                requested_regions: requestedRegions
-            }
-        ).then((response) => {
-            if (response.errors.length) {
-                handleErrors(response.errors);
-                return;
-            }
-            setProposalDataState(response.data.proposalData);
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
+        if (modalState == "confirm_offer_trade") {
+            apiFetch(
+                villageAPI,
+                {
+                    request: 'CreateProposal',
+                    type: 'offer_trade',
+                    target_village_id: strategicDisplayRight.village.village_id,
+                    offered_resources: offeredResources,
+                    offered_regions: offeredRegions,
+                    requested_resources: requestedResources,
+                    requested_regions: requestedRegions
+                }
+            ).then((response) => {
+                if (response.errors.length) {
+                    handleErrors(response.errors);
+                    return;
+                }
+                setProposalDataState(response.data.proposalData);
+                setModalText(response.data.response_message);
+                setModalHeader("Confirmation");
+                setModalState("response_message");
             });
-        });
+        }
+        else {
+            setModalState("confirm_offer_trade");
+            setModalText(null);
+            setModalHeader("Trade resources and regions");
+        }
     }
     const EnactProposal = () => {
-        apiFetch(
-            villageAPI,
-            {
-                request: 'EnactProposal',
-                proposal_id: currentProposal.proposal_id,
-            }
-        ).then((response) => {
-            if (response.errors.length) {
-                handleErrors(response.errors);
-                return;
-            }
-            setProposalDataState(response.data.proposalData);
-            setCurrentProposal(null);
-            setCurrentProposalKey(null);
-            setPolicyDataState(response.data.policyData);
-            setDisplayPolicyID(response.data.policyData.policy_id);
-            setPolicyDisplay(getPolicyDisplayData(response.data.policyData.policy_id));
-            setStrategicDataState(response.data.strategicData);
-            setStrategicDisplayLeft(response.data.strategicData.find(item => item.village.name == villageName));
-            setStrategicDisplayRight(response.data.strategicData.find(item => item.village.name == strategicDisplayRight.village.name));
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
+        if (modalState == "confirm_enact_proposal") {
+            apiFetch(
+                villageAPI,
+                {
+                    request: 'EnactProposal',
+                    proposal_id: currentProposal.proposal_id,
+                }
+            ).then((response) => {
+                if (response.errors.length) {
+                    handleErrors(response.errors);
+                    return;
+                }
+                setProposalDataState(response.data.proposalData);
+                setCurrentProposal(null);
+                setCurrentProposalKey(null);
+                setPolicyDataState(response.data.policyData);
+                setDisplayPolicyID(response.data.policyData.policy_id);
+                setPolicyDisplay(getPolicyDisplayData(response.data.policyData.policy_id));
+                setStrategicDataState(response.data.strategicData);
+                setStrategicDisplayLeft(response.data.strategicData.find(item => item.village.name == villageName));
+                setStrategicDisplayRight(response.data.strategicData.find(item => item.village.name == strategicDisplayRight.village.name));
+                setModalText(response.data.response_message);
+                setModalHeader("Confirmation");
+                setModalState("response_message");
             });
-        });
+        }
+        else {
+            setModalState("confirm_enact_proposal");
+            setModalText("Are you sure you want to enact this proposal?");
+            setModalHeader("Confirmation");
+        }
     }
     const BoostVote = () => {
-        apiFetch(
-            villageAPI,
-            {
-                request: 'BoostVote',
-                proposal_id: currentProposal.proposal_id,
-            }
-        ).then((response) => {
-            if (response.errors.length) {
-                handleErrors(response.errors);
-                return;
-            }
-            setProposalDataState(response.data.proposalData);
-            setCurrentProposal(response.data.proposalData[currentProposalKey]);
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
+        if (modalState == "confirm_boost_vote") {
+            apiFetch(
+                villageAPI,
+                {
+                    request: 'BoostVote',
+                    proposal_id: currentProposal.proposal_id,
+                }
+            ).then((response) => {
+                if (response.errors.length) {
+                    handleErrors(response.errors);
+                    return;
+                }
+                setProposalDataState(response.data.proposalData);
+                setCurrentProposal(response.data.proposalData[currentProposalKey]);
+                setModalText(response.data.response_message);
+                setModalHeader("Confirmation");
+                setModalState("response_message");
             });
-        });
+        }
+        else {
+            setModalState("confirm_boost_vote");
+            setModalText("When a vote Against is boosted:\n The Kage will lose 500 Reputation when the proposal is enacted.\n\nWhen a vote In Favor is boosted:\nTotal Reputation loss from Against votes will be reduced by 500.\n\nBoosting a vote will cost 500 Reputation when the proposal is passed.\n\nHowever, a boosted vote In Favor will only cost Reputation if there is a boosted vote Against. If there are more boosted votes In Favor than Against, the cost will be split between between votes In Favor.");
+            setModalHeader("Confirmation");
+        }
     }
     const CancelVote = () => {
-        apiFetch(
-            villageAPI,
-            {
-                request: 'CancelVote',
-                proposal_id: currentProposal.proposal_id,
-            }
-        ).then((response) => {
-            if (response.errors.length) {
-                handleErrors(response.errors);
-                return;
-            }
-            setProposalDataState(response.data.proposalData);
-            setCurrentProposal(response.data.proposalData[currentProposalKey]);
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
+        if (modalState == "confirm_cancel_vote") {
+            apiFetch(
+                villageAPI,
+                {
+                    request: 'CancelVote',
+                    proposal_id: currentProposal.proposal_id,
+                }
+            ).then((response) => {
+                if (response.errors.length) {
+                    handleErrors(response.errors);
+                    return;
+                }
+                setProposalDataState(response.data.proposalData);
+                setCurrentProposal(response.data.proposalData[currentProposalKey]);
+                setModalText(response.data.response_message);
+                setModalHeader("Confirmation");
+                setModalState("response_message");
             });
-        });
+        }
+        else {
+            setModalState("confirm_cancel_vote");
+            setModalText("Are you sure you wish to cancel your vote for this proposal?");
+            setModalHeader("Confirmation");
+        }
     }
     const SubmitVote = (vote) => {
         apiFetch(
@@ -296,6 +336,11 @@ export function KageQuarters({
             setCurrentProposal(response.data.proposalData[currentProposalKey]);
         });
     }
+    const ViewTrade = () => {
+        setModalState("view_trade");
+        setModalText(null);
+        setModalHeader("View trade offer");
+    }
     React.useEffect(() => {
         if (proposalDataState.length && currentProposal === null) {
             setCurrentProposal(proposalDataState[0]);
@@ -305,6 +350,117 @@ export function KageQuarters({
     }, [proposalDataState]);
     return (
         <>
+            {modalState !== "closed" &&
+                <>
+                    <div className="modal_backdrop"></div>
+                    <div className="modal">
+                        <div className="modal_header">{modalHeader}</div>
+                        <div className="modal_text">{modalText}</div>
+                        {modalState == "confirm_policy" &&
+                            <>
+                                <div className="modal_confirm_button" onClick={() => ChangePolicy()}>Confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "confirm_cancel_proposal" &&
+                            <>
+                                <div className="modal_confirm_button" onClick={() => CancelProposal()}>Confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "confirm_enact_proposal" &&
+                            <>
+                                <div className="modal_confirm_button" onClick={() => EnactProposal()}>Confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "confirm_boost_vote" &&
+                            <>
+                                <div className="modal_confirm_button" onClick={() => BoostVote()}>Confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "confirm_cancel_vote" &&
+                            <>
+                                <div className="modal_confirm_button" onClick={() => CancelVote()}>Confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "confirm_declare_war" &&
+                            <>
+                                <div className="modal_confirm_button" onClick={() => DeclareWar()}>Confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "confirm_offer_peace" &&
+                            <>
+                                <div className="modal_confirm_button" onClick={() => OfferPeace()}>Confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "confirm_form_alliance" &&
+                            <>
+                                <div className="modal_confirm_button" onClick={() => OfferAlliance()}>Confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "confirm_break_alliance" &&
+                            <>
+                                <div className="modal_confirm_button" onClick={() => BreakAlliance()}>Confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "confirm_offer_trade" &&
+                            <>
+                                <div class="schedule_challenge_subtext_wrapper" style={{ marginBottom: "20px", marginTop: "-10px" }}>
+                                    <span class="schedule_challenge_subtext">Each village can offer up to 25000 resources of each resource type per trade.</span>
+                                    <span class="schedule_challenge_subtext">Trades have a cooldown of 24 hours.</span>
+                                </div>
+                                {TradeDisplay({
+                                    viewOnly: false,
+                                    offeringVillageResources: resourceDataState,
+                                    offeringVillageRegions: strategicDisplayLeft.regions,
+                                    offeredResources: offeredResources,
+                                    setOfferedResources: setOfferedResources,
+                                    offeredRegions: offeredRegions,
+                                    setOfferedRegions: setOfferedRegions,
+                                    targetVillageResources: null,
+                                    targetVillageRegions: strategicDisplayRight.regions,
+                                    requestedResources: requestedResources,
+                                    setRequestedResources: setRequestedResources,
+                                    requestedRegions: requestedRegions,
+                                    setRequestedRegions: setRequestedRegions,
+                                })}
+                                <div className="modal_confirm_button" onClick={() => OfferTrade()}>confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "view_trade" &&
+                            <>
+                                <TradeDisplay
+                                    viewOnly={true}
+                                    offeringVillageResources={null}
+                                    offeringVillageRegions={null}
+                                    offeredResources={currentProposal.trade_data.offered_resources}
+                                    setOfferedResources={null}
+                                    offeredRegions={currentProposal.trade_data.offered_regions}
+                                    setOfferedRegions={null}
+                                    targetVillageResources={null}
+                                    targetVillageRegions={null}
+                                    requestedResources={currentProposal.trade_data.requested_resources}
+                                    setRequestedResources={null}
+                                    requestedRegions={currentProposal.trade_data.requested_regions}
+                                    setRequestedRegions={null}
+                                />
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>close</div>
+                            </>
+                        }
+                        {modalState == "response_message" &&
+                            <div className="modal_close_button" onClick={() => setModalState("closed")}>close</div>
+                        }
+                    </div>
+                </>
+            }
             <div className="kq_container">
                 <div className="row first">
                     <div className="column first">
@@ -387,38 +543,10 @@ export function KageQuarters({
                                     {playerSeatState.seat_type == "kage" &&
                                         <>
                                             <div className="proposal_cancel_button_wrapper">
-                                            <div className={currentProposal ? "proposal_cancel_button" : "proposal_cancel_button disabled"} onClick={() => openModal({
-                                                header: 'Confirmation',
-                                                text: "Are you sure you want to cancel this proposal?",
-                                                ContentComponent: null,
-                                                onConfirm: () => CancelProposal(),
-                                            })}>cancel proposal</div>
+                                                <div className={currentProposal ? "proposal_cancel_button" : "proposal_cancel_button disabled"} onClick={() => CancelProposal()}>cancel proposal</div>
                                             </div>
                                             {(currentProposal && (currentProposal.type == "offer_trade" || currentProposal.type == "accept_trade")) &&
-                                                <div className="trade_view_button_wrapper alliance"
-                                                onClick={
-                                                    () => openModal({
-                                                        header: 'View trade offer',
-                                                        text: '',
-                                                        ContentComponent: TradeDisplay,
-                                                        componentProps: ({
-                                                            viewOnly: true,
-                                                            offeringVillageResources: resourceDataState,
-                                                            offeringVillageRegions: strategicDisplayLeft.regions,
-                                                            offeredResources: offeredResources,
-                                                            setOfferedResources: setOfferedResources,
-                                                            offeredRegions: offeredRegions,
-                                                            setOfferedRegions: setOfferedRegions,
-                                                            targetVillageResources: null,
-                                                            targetVillageRegions: strategicDisplayRight.regions,
-                                                            requestedResources: requestedResources,
-                                                            setRequestedResources: setRequestedResources,
-                                                            requestedRegions: requestedRegions,
-                                                            setRequestedRegions: setRequestedRegions,
-                                                        }),
-                                                        onConfirm: null,
-                                                    })
-                                                }>
+                                                <div className="trade_view_button_wrapper alliance" onClick={() => ViewTrade()}>
                                                     <div className="trade_view_button_inner">
                                                         <img src="/images/v2/icons/trade.png" className="trade_view_button_icon" />
                                                     </div>
@@ -427,12 +555,7 @@ export function KageQuarters({
                                             <div className="proposal_enact_button_wrapper">
                                                 <div className={(currentProposal && (currentProposal.enact_time_remaining !== null ||
                                                     currentProposal.votes.length == seatDataState.filter(seat => seat.seat_type == "elder" && seat.seat_id != null).length
-                                            )) ? "proposal_enact_button" : "proposal_enact_button disabled"} onClick={() => openModal({
-                                                header: 'Confirmation',
-                                                text: 'Are you sure you want to enact this proposal?',
-                                                ContentComponent: null,
-                                                onConfirm: () => EnactProposal(),
-                                            })}>enact proposal</div>
+                                                )) ? "proposal_enact_button" : "proposal_enact_button disabled"} onClick={() => EnactProposal()}>enact proposal</div>
                                                 {/*proposalRepAdjustment > 0 &&
                                                 <div className="rep_change positive">REPUATION GAIN: +{proposalRepAdjustment}</div>
                                             */}
@@ -460,30 +583,7 @@ export function KageQuarters({
                                                         <div className="proposal_yes_button" onClick={() => SubmitVote(1)}>vote in favor</div>
                                                     </div>
                                                     {(currentProposal && (currentProposal.type == "offer_trade" || currentProposal.type == "accept_trade")) &&
-                                                        <div className="trade_view_button_wrapper alliance"
-                                                            onClick={
-                                                                () => openModal({
-                                                                    header: 'View trade offer',
-                                                                    text: '',
-                                                                    ContentComponent: TradeDisplay,
-                                                                    componentProps: ({
-                                                                        viewOnly: true,
-                                                                        offeringVillageResources: resourceDataState,
-                                                                        offeringVillageRegions: strategicDisplayLeft.regions,
-                                                                        offeredResources: offeredResources,
-                                                                        setOfferedResources: setOfferedResources,
-                                                                        offeredRegions: offeredRegions,
-                                                                        setOfferedRegions: setOfferedRegions,
-                                                                        targetVillageResources: null,
-                                                                        targetVillageRegions: strategicDisplayRight.regions,
-                                                                        requestedResources: requestedResources,
-                                                                        setRequestedResources: setRequestedResources,
-                                                                        requestedRegions: requestedRegions,
-                                                                        setRequestedRegions: setRequestedRegions,
-                                                                    }),
-                                                                    onConfirm: null,
-                                                                })
-                                                            }>
+                                                        <div className="trade_view_button_wrapper alliance" onClick={() => ViewTrade()}>
                                                             <div className="trade_view_button_inner">
                                                                 <img src="/images/v2/icons/trade.png" className="trade_view_button_icon" />
                                                             </div>
@@ -500,30 +600,7 @@ export function KageQuarters({
                                                         <div className="proposal_yes_button disabled">vote in favor</div>
                                                     </div>
                                                     {(currentProposal && (currentProposal.type == "offer_trade" || currentProposal.type == "accept_trade")) &&
-                                                        <div className="trade_view_button_wrapper alliance"
-                                                            onClick={
-                                                                () => openModal({
-                                                                    header: 'View trade offer',
-                                                                    text: '',
-                                                                    ContentComponent: TradeDisplay,
-                                                                    componentProps: ({
-                                                                        viewOnly: true,
-                                                                        offeringVillageResources: resourceDataState,
-                                                                        offeringVillageRegions: strategicDisplayLeft.regions,
-                                                                        offeredResources: offeredResources,
-                                                                        setOfferedResources: setOfferedResources,
-                                                                        offeredRegions: offeredRegions,
-                                                                        setOfferedRegions: setOfferedRegions,
-                                                                        targetVillageResources: null,
-                                                                        targetVillageRegions: strategicDisplayRight.regions,
-                                                                        requestedResources: requestedResources,
-                                                                        setRequestedResources: setRequestedResources,
-                                                                        requestedRegions: requestedRegions,
-                                                                        setRequestedRegions: setRequestedRegions,
-                                                                    }),
-                                                                    onConfirm: null,
-                                                                })
-                                                            }>
+                                                        <div className="trade_view_button_wrapper alliance" onClick={() => ViewTrade()}>
                                                             <div className="trade_view_button_inner">
                                                                 <img src="/images/v2/icons/trade.png" className="trade_view_button_icon" />
                                                             </div>
@@ -537,42 +614,10 @@ export function KageQuarters({
                                             {(currentProposal && currentProposal.vote_time_remaining != null && currentProposal.votes.find(vote => vote.user_id == playerID)) &&
                                                 <>
                                                     <div className="proposal_cancel_vote_button_wrapper">
-                                                        <div className="proposal_cancel_vote_button" 
-                                                            onClick={
-                                                                () => openModal({
-                                                                    header: 'Confirmation',
-                                                                    text: null,
-                                                                    ContentComponent: null,
-                                                                    onConfirm: () => CancelVote(),
-                                                                })
-                                                            }>
-                                                            change vote</div>
+                                                        <div className="proposal_cancel_vote_button" onClick={() => CancelVote()}>change vote</div>
                                                     </div>
                                                     {(currentProposal && (currentProposal.type == "offer_trade" || currentProposal.type == "accept_trade")) &&
-                                                        <div className="trade_view_button_wrapper alliance"
-                                                            onClick={
-                                                                () => openModal({
-                                                                    header: 'View trade offer',
-                                                                    text: '',
-                                                                    ContentComponent: TradeDisplay,
-                                                                    componentProps: ({
-                                                                        viewOnly: true,
-                                                                        offeringVillageResources: resourceDataState,
-                                                                        offeringVillageRegions: strategicDisplayLeft.regions,
-                                                                        offeredResources: offeredResources,
-                                                                        setOfferedResources: setOfferedResources,
-                                                                        offeredRegions: offeredRegions,
-                                                                        setOfferedRegions: setOfferedRegions,
-                                                                        targetVillageResources: null,
-                                                                        targetVillageRegions: strategicDisplayRight.regions,
-                                                                        requestedResources: requestedResources,
-                                                                        setRequestedResources: setRequestedResources,
-                                                                        requestedRegions: requestedRegions,
-                                                                        setRequestedRegions: setRequestedRegions,
-                                                                    }),
-                                                                    onConfirm: null,
-                                                                })
-                                                            }>
+                                                        <div className="trade_view_button_wrapper alliance" onClick={() => ViewTrade()}>
                                                             <div className="trade_view_button_inner">
                                                                 <img src="/images/v2/icons/trade.png" className="trade_view_button_icon" />
                                                             </div>
@@ -580,12 +625,7 @@ export function KageQuarters({
                                                     }
                                                     {(currentProposal.votes.find(vote => vote.user_id == playerID).rep_adjustment == 0) &&
                                                         <div className="proposal_boost_vote_button_wrapper">
-                                                    <div className="proposal_boost_vote_button" onClick={() => openModal({
-                                                        header: 'Confirmation',
-                                                        text: "When a vote Against is boosted:\n The Kage will lose 500 Reputation when the proposal is enacted.\n\nWhen a vote In Favor is boosted:\nTotal Reputation loss from Against votes will be reduced by 500.\n\nBoosting a vote will cost 500 Reputation when the proposal is passed.\n\nHowever, a boosted vote In Favor will only cost Reputation if there is a boosted vote Against. If there are more boosted votes In Favor than Against, the cost will be split between between votes In Favor.",
-                                                        ContentComponent: null,
-                                                        onConfirm: () => BoostVote(),
-                                                    })}>boost vote</div>
+                                                            <div className="proposal_boost_vote_button" onClick={() => BoostVote()}>boost vote</div>
                                                         </div>
                                                     }
                                                 </>
@@ -596,30 +636,7 @@ export function KageQuarters({
                                                         <div className="proposal_cancel_vote_button disabled">cancel vote</div>
                                                     </div>
                                                     {(currentProposal && (currentProposal.type == "offer_trade" || currentProposal.type == "accept_trade")) &&
-                                                        <div className="trade_view_button_wrapper alliance"
-                                                            onClick={
-                                                                () => openModal({
-                                                                    header: 'View trade offer',
-                                                                    text: '',
-                                                                    ContentComponent: TradeDisplay,
-                                                                    componentProps: ({
-                                                                        viewOnly: true,
-                                                                        offeringVillageResources: resourceDataState,
-                                                                        offeringVillageRegions: strategicDisplayLeft.regions,
-                                                                        offeredResources: offeredResources,
-                                                                        setOfferedResources: setOfferedResources,
-                                                                        offeredRegions: offeredRegions,
-                                                                        setOfferedRegions: setOfferedRegions,
-                                                                        targetVillageResources: null,
-                                                                        targetVillageRegions: strategicDisplayRight.regions,
-                                                                        requestedResources: requestedResources,
-                                                                        setRequestedResources: setRequestedResources,
-                                                                        requestedRegions: requestedRegions,
-                                                                        setRequestedRegions: setRequestedRegions,
-                                                                    }),
-                                                                    onConfirm: null,
-                                                                })
-                                                            }>
+                                                        <div className="trade_view_button_wrapper alliance" onClick={() => ViewTrade()}>
                                                             <div className="trade_view_button_inner">
                                                                 <img src="/images/v2/icons/trade.png" className="trade_view_button_icon" />
                                                             </div>
@@ -711,12 +728,7 @@ export function KageQuarters({
                                 ))}
                             </div>
                             {displayPolicyID != policyDataState.policy_id &&
-                                <div className={playerSeatState.seat_type == "kage" ? "village_policy_change_button" : "village_policy_change_button disabled"} onClick={() => openModal({
-                                    header: 'Confirmation',
-                                    text: "Are you sure you want to change policies? You will be unable to select a new policy for 3 days.",
-                                    ContentComponent: null,
-                                    onConfirm: () => ChangePolicy(),
-                                })}>change policy</div>
+                                <div className={playerSeatState.seat_type == "kage" ? "village_policy_change_button" : "village_policy_change_button disabled"} onClick={() => ChangePolicy()}>change policy</div>
                             }
                             <div className="village_policy_main_container">
                                 <div className="village_policy_main_inner">
@@ -745,6 +757,15 @@ export function KageQuarters({
                                 </div>
                             </div>
                             <div className="village_policy_penalty_container">
+                                {policyDisplay.resources.map((resource, index) => (
+                                    <div key={index} className="policy_resource_item">
+                                        <svg width="16" height="16" viewBox="0 0 100 100">
+                                            <polygon points="25,20 50,45 25,70 0,45" fill="#414b8c" />
+                                            <polygon points="25,0 50,25 25,50 0,25" fill="#5964a6" />
+                                        </svg>
+                                        <div className="policy_resource_text">{resource}</div>
+                                    </div>
+                                ))}
                                 {policyDisplay.penalties.map((penalty, index) => (
                                     <div key={index} className="policy_penalty_item">
                                         <svg width="16" height="16" viewBox="0 0 100 100">
@@ -775,44 +796,24 @@ export function KageQuarters({
                             <div className="strategic_info_navigation">
                                 <div className="strategic_info_navigation_diplomacy_buttons">
                                     {strategicDisplayLeft.enemies.find(enemy => enemy == strategicDisplayRight.village.name) ?
-                                        <div className="diplomacy_action_button_wrapper war cancel" onClick={() => openModal({
-                                            header: 'Confirmation',
-                                            text: "Are you sure you want to offer peace with " + strategicDisplayRight.village.name + "?",
-                                            ContentComponent: null,
-                                            onConfirm: () => OfferPeace(),
-                                        })}>
+                                        <div className="diplomacy_action_button_wrapper war cancel" onClick={() => OfferPeace()}>
                                             <div className="diplomacy_action_button_inner">
                                             </div>
                                         </div>
                                         :
-                                        <div className="diplomacy_action_button_wrapper war" onClick={() => openModal({
-                                            header: 'Confirmation',
-                                            text: "Are you sure you declare war with " + strategicDisplayRight.village.name + "?",
-                                            ContentComponent: null,
-                                            onConfirm: () => DeclareWar(),
-                                        })}>
+                                        <div className="diplomacy_action_button_wrapper war" onClick={() => DeclareWar()}>
                                             <div className="diplomacy_action_button_inner">
                                                 <img src="/images/icons/war.png" className="diplomacy_action_button_icon" />
                                             </div>
                                         </div>
                                     }
                                     {strategicDisplayLeft.allies.find(ally => ally == strategicDisplayRight.village.name) ?
-                                        <div className="diplomacy_action_button_wrapper alliance cancel" onClick={() => openModal({
-                                            header: 'Confirmation',
-                                            text: "Are you sure you want break an alliance with " + strategicDisplayRight.village.name + "?",
-                                            ContentComponent: null,
-                                            onConfirm: () => BreakAlliance(),
-                                        })}>
+                                        <div className="diplomacy_action_button_wrapper alliance cancel" onClick={() => BreakAlliance()}>
                                             <div className="diplomacy_action_button_inner">
                                             </div>
                                         </div>
                                         :
-                                        <div className="diplomacy_action_button_wrapper alliance" onClick={() => openModal({
-                                            header: 'Confirmation',
-                                            text: "Are you sure you want to form an alliance with " + strategicDisplayRight.village.name + "?\nYou can be a member of only one Alliance at any given time.",
-                                            ContentComponent: null,
-                                            onConfirm: () => OfferAlliance(),
-                                        })}>
+                                        <div className="diplomacy_action_button_wrapper alliance" onClick={() => OfferAlliance()}>
                                             <div className="diplomacy_action_button_inner">
                                                 <img src="/images/icons/ally.png" className="diplomacy_action_button_icon" />
                                             </div>
@@ -858,30 +859,7 @@ export function KageQuarters({
                                 </div>
                                 <div className="strategic_info_navigation_diplomacy_buttons">
                                     {strategicDisplayLeft.allies.find(ally => ally == strategicDisplayRight.village.name) &&
-                                        <div className="diplomacy_action_button_wrapper alliance"
-                                            onClick={
-                                                () => openModal({
-                                                    header: 'View trade offer',
-                                                    text: '',
-                                                    ContentComponent: TradeDisplay,
-                                                    componentProps: ({
-                                                        viewOnly: false,
-                                                        offeringVillageResources: resourceDataState,
-                                                        offeringVillageRegions: strategicDisplayLeft.regions,
-                                                        offeredResources: offeredResources,
-                                                        setOfferedResources: setOfferedResources,
-                                                        offeredRegions: offeredRegions,
-                                                        setOfferedRegions: setOfferedRegions,
-                                                        targetVillageResources: null,
-                                                        targetVillageRegions: strategicDisplayRight.regions,
-                                                        requestedResources: requestedResources,
-                                                        setRequestedResources: setRequestedResources,
-                                                        requestedRegions: requestedRegions,
-                                                        setRequestedRegions: setRequestedRegions,
-                                                    }),
-                                                    onConfirm: () => OfferTrade(),
-                                                })
-                                            }>
+                                        <div className="diplomacy_action_button_wrapper alliance" onClick={() => OfferTrade()}>
                                             <div className="diplomacy_action_button_inner">
                                                 <img src="/images/v2/icons/trade.png" className="diplomacy_action_button_icon" />
                                             </div>
@@ -899,6 +877,243 @@ export function KageQuarters({
             </div>
         </>
     );
+
+    function TradeDisplay({
+        viewOnly,
+        offeringVillageResources,
+        offeringVillageRegions,
+        offeredResources,
+        setOfferedResources,
+        offeredRegions,
+        setOfferedRegions,
+        targetVillageResources,
+        targetVillageRegions,
+        requestedResources,
+        setRequestedResources,
+        requestedRegions,
+        setRequestedRegions,
+    }) {
+        const toggleOfferedRegion = (regionId) => {
+            setOfferedRegions(current => {
+                // Check if the region is already selected
+                if (current.includes(regionId)) {
+                    // If it is, filter it out (unselect it)
+                    return current.filter(id => id !== regionId);
+                } else {
+                    // Otherwise, add it to the selected regions
+                    return [...current, regionId];
+                }
+            });
+        };
+        const toggleRequestedRegion = (regionId) => {
+            setRequestedRegions(current => {
+                // Check if the region is already selected
+                if (current.includes(regionId)) {
+                    // If it is, filter it out (unselect it)
+                    return current.filter(id => id !== regionId);
+                } else {
+                    // Otherwise, add it to the selected regions
+                    return [...current, regionId];
+                }
+            });
+        };
+        const handleOfferedResourcesChange = (resourceName, value) => {
+            setOfferedResources(currentResources =>
+                currentResources.map(resource =>
+                    resource.resource_name === resourceName
+                        ? { ...resource, count: value }
+                        : resource
+                )
+            );
+        };
+        const handleRequestedResourcesChange = (resourceName, value) => {
+            setRequestedResources(currentResources =>
+                currentResources.map(resource =>
+                    resource.resource_name === resourceName
+                        ? { ...resource, count: value }
+                        : resource
+                )
+            );
+        };
+
+        return (
+            viewOnly ?
+                <div className="trade_display_container">
+                    <div className="trade_display_offer_container">
+                        <div className="header">Offered Resources</div>
+                        <div className="trade_display_resources">
+                            {offeredResources
+                                .map((resource, index) => {
+                                    const total = offeringVillageResources ? offeringVillageResources.find(total => total.resource_id === resource.resource_id).count : null;
+                                    return (
+                                        <div key={resource.resource_id} className="trade_display_resource_wrapper">
+                                            <input
+                                                type="text"
+                                                min="0"
+                                                max={total ? total : 25000}
+                                                step="100"
+                                                placeholder="0"
+                                                className="trade_display_resource_input"
+                                                value={resource.count}
+                                                style={{ userSelect: "none" }}
+                                                readOnly
+                                            />
+                                            <div className="trade_display_resource_name">{resource.resource_name}</div>
+                                            {total ?
+                                                <div className="trade_display_resource_total">{total}</div>
+                                                :
+                                                <div className="trade_display_resource_total">???</div>
+                                            }
+                                        </div>
+                                    );
+                                })}
+                        </div>
+                        <div className="header">Offered Regions</div>
+                        <div className="trade_display_regions">
+                            {offeredRegions
+                                .filter(region => region.region_id > 5)
+                                .map((region, index) => (
+                                    <div key={region.name} className="trade_display_region_wrapper">
+                                        <div className="trade_display_region_name">{region.name}</div>
+                                    </div>
+                                ))}
+                        </div>
+                    </div>
+                    <div className="trade_display_request_container">
+                        <div className="header">Requested Resources</div>
+                        <div className="trade_display_resources">
+                            {requestedResources
+                                .map((resource, index) => {
+                                    const total = targetVillageResources ? targetVillageResources.find(total => total.resource_id === resource.resource_id).count : null;
+                                    return (
+                                        <div key={resource.resource_id} className="trade_display_resource_wrapper">
+                                            <input
+                                                type="text"
+                                                min="0"
+                                                max={total ? total : 25000}
+                                                step="100"
+                                                placeholder="0"
+                                                className="trade_display_resource_input"
+                                                value={resource.count}
+                                                onChange={(e) => handleRequestedResourcesChange(resource.resource_name, parseInt(e.target.value))}
+                                                style={{ userSelect: "none" }}
+                                                readOnly
+                                            />
+                                            <div className="trade_display_resource_name">{resource.resource_name}</div>
+                                            {total ?
+                                                <div className="trade_display_resource_total">{total}</div>
+                                                :
+                                                <div className="trade_display_resource_total">???</div>
+                                            }
+                                        </div>
+                                    );
+                                })}
+                        </div>
+                        <div className="header">Requested Regions</div>
+                        <div className="trade_display_regions">
+                            {requestedRegions
+                                .filter(region => region.region_id > 5)
+                                .map((region, index) => (
+                                    <div key={region.name} className="trade_display_region_wrapper">
+                                        <div className="trade_display_region_name">{region.name}</div>
+                                    </div>
+                                ))}
+                        </div>
+                    </div>
+                </div>
+                :
+                <div className="trade_display_container">
+                    <div className="trade_display_offer_container">
+                        <div className="header">Offer Resources</div>
+                        <div className="trade_display_resources">
+                            {offeredResources
+                                .map((resource, index) => {
+                                    const total = offeringVillageResources ? offeringVillageResources.find(total => total.resource_id === resource.resource_id).count : null;
+                                    return (
+                                        <div key={resource.resource_id} className="trade_display_resource_wrapper">
+                                            <input
+                                                type="number"
+                                                min="0"
+                                                max={total ? total : 25000}
+                                                step="100"
+                                                placeholder="0"
+                                                className="trade_display_resource_input"
+                                                value={resource.count}
+                                                onChange={(e) => handleOfferedResourcesChange(resource.resource_name, parseInt(e.target.value))}
+                                            />
+                                            <div className="trade_display_resource_name">{resource.resource_name}</div>
+                                            {total ?
+                                                <div className="trade_display_resource_total">{total}</div>
+                                                :
+                                                <div className="trade_display_resource_total">???</div>
+                                            }
+                                        </div>
+                                    );
+                                })}
+                        </div>
+                        <div className="header">Offer Regions</div>
+                        <div className="trade_display_regions">
+                            {offeringVillageRegions
+                                .filter(region => region.region_id > 5)
+                                .map((region, index) => (
+                                    <div key={region.name} className="trade_display_region_wrapper">
+                                        <div className="trade_display_region_name">{region.name}</div>
+                                        <input
+                                            type="checkbox"
+                                            checked={offeredRegions.includes(region.region_id)}
+                                            onChange={() => toggleOfferedRegion(region.region_id)}
+                                        />
+                                    </div>
+                                ))}
+                        </div>
+                    </div>
+                    <div className="trade_display_request_container">
+                        <div className="header">Request Resources</div>
+                        <div className="trade_display_resources">
+                            {requestedResources
+                                .map((resource, index) => {
+                                    const total = targetVillageResources ? targetVillageResources.find(total => total.resource_id === resource.resource_id).count : null;
+                                    return (
+                                        <div key={resource.resource_id} className="trade_display_resource_wrapper">
+                                            <input
+                                                type="number"
+                                                min="0"
+                                                max={total ? total : 25000}
+                                                step="100"
+                                                placeholder="0"
+                                                className="trade_display_resource_input"
+                                                value={resource.count}
+                                                onChange={(e) => handleRequestedResourcesChange(resource.resource_name, parseInt(e.target.value))}
+                                            />
+                                            <div className="trade_display_resource_name">{resource.resource_name}</div>
+                                            {total ?
+                                                <div className="trade_display_resource_total">{total}</div>
+                                                :
+                                                <div className="trade_display_resource_total">???</div>
+                                            }
+                                        </div>
+                                    );
+                                })}
+                        </div>
+                        <div className="header">Request Regions</div>
+                        <div className="trade_display_regions">
+                            {targetVillageRegions
+                                .filter(region => region.region_id > 5)
+                                .map((region, index) => (
+                                    <div key={region.name} className="trade_display_region_wrapper">
+                                        <div className="trade_display_region_name">{region.name}</div>
+                                        <input
+                                            type="checkbox"
+                                            checked={requestedRegions.includes(region.region_id)}
+                                            onChange={() => toggleRequestedRegion(region.region_id)}
+                                        />
+                                    </div>
+                                ))}
+                        </div>
+                    </div>
+                </div>
+        );
+    }
 
     function cyclePolicy(direction) {
         var newPolicyID;
