+++ ui_components/src/village/VillageHQ.js
@@ -1,5 +1,5 @@
 import { apiFetch } from "../utils/network.js";
-import { useModal } from '../utils/modalContext.js';
+
 export function VillageHQ({
     playerID,
     playerSeatState,
@@ -23,12 +23,14 @@ export function VillageHQ({
     getVillageIcon,
     getPolicyDisplayData
 }) {
+    const [modalState, setModalState] = React.useState("closed");
     const [resourceDaysToShow, setResourceDaysToShow] = React.useState(1);
     const [policyDisplay, setPolicyDisplay] = React.useState(getPolicyDisplayData(policyDataState.policy_id));
-    const selectedTimesUTC = React.useRef([]);
-    const selectedTimeUTC = React.useRef(null);
-    const challengeTarget = React.useRef(null);
-    const { openModal } = useModal();
+    const [selectedTimesUTC, setSelectedTimesUTC] = React.useState([]);
+    const [selectedTimeUTC, setSelectedTimeUTC] = React.useState(null);
+    const [modalHeader, setModalHeader] = React.useState(null);
+    const [modalText, setModalText] = React.useState(null);
+    const [challengeTarget, setChallengeTarget] = React.useState(null);
     const DisplayFromDays = (days) => {
         switch (days) {
             case 1:
@@ -87,74 +89,52 @@ export function VillageHQ({
             }
             setSeatDataState(response.data.seatData);
             setPlayerSeatState(response.data.playerSeat);
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
-            });
+            setModalHeader("Confirmation");
+            setModalText(response.data.response_message);
+            setModalState("response_message");
         });
     }
     const Resign = () => {
-        apiFetch(
-            villageAPI,
-            {
-                request: 'Resign',
-            }
-        ).then((response) => {
-            if (response.errors.length) {
-                handleErrors(response.errors);
-                return;
-            }
-            setSeatDataState(response.data.seatData);
-            setPlayerSeatState(response.data.playerSeat);
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
+        if (modalState == "confirm_resign") {
+            apiFetch(
+                villageAPI,
+                {
+                    request: 'Resign',
+                }
+            ).then((response) => {
+                if (response.errors.length) {
+                    handleErrors(response.errors);
+                    return;
+                }
+                setSeatDataState(response.data.seatData);
+                setPlayerSeatState(response.data.playerSeat);
+                setModalHeader("Confirmation");
+                setModalText(response.data.response_message);
+                setModalState("response_message");
             });
-        });
-    }
-    const [challengeSubmittedFlag, setChallengeSubmittedFlag] = React.useState(false);
-    React.useEffect(() => {
-        if (challengeSubmittedFlag) {
-            ConfirmSubmitChallenge();
         }
-    }, [challengeSubmittedFlag]);
+        else {
+            setModalHeader("Confirmation");
+            setModalState("confirm_resign");
+            setModalText("Are you sure you wish to resign from your current position?");
+        }
+    }
     const Challenge = (target_seat) => {
-        challengeTarget.current = target_seat;
-        openModal({
-            header: 'Submit Challenge',
-            text: "Select times below that you are available to battle.",
-            ContentComponent: TimeGrid,
-            componentProps: ({
-                selectedTimesUTC: selectedTimesUTC,
-                startHourUTC: luxon.DateTime.fromObject({ hour: 0, zone: luxon.Settings.defaultZoneName }).toUTC().hour
-            }),
-            onConfirm: () => setChallengeSubmittedFlag(true),
-        });
+        setChallengeTarget(target_seat);
+        setModalState("submit_challenge");
+        setModalHeader("Submit Challenge");
+        setModalText("Select times below that you are available to battle.");
     }
     const ConfirmSubmitChallenge = () => {
-        setChallengeSubmittedFlag(false);
         if (selectedTimesUTC.length < 12) {
-            openModal({
-                header: 'Submit Challenge',
-                text: "You must select at least 12 slots.",
-                ContentComponent: TimeGrid,
-                componentProps: ({
-                    selectedTimesUTC: selectedTimesUTC,
-                    startHourUTC: luxon.DateTime.fromObject({ hour: 0, zone: luxon.Settings.defaultZoneName }).toUTC().hour
-                }),
-                onConfirm: () => setChallengeSubmittedFlag(true),
-            });
+            setModalText("Insufficient slots selected.");
         } else {
             apiFetch(
                 villageAPI,
                 {
                     request: 'SubmitChallenge',
-                    seat_id: challengeTarget.current.seat_id,
-                    selected_times: selectedTimesUTC.current,
+                    seat_id: challengeTarget.seat_id,
+                    selected_times: selectedTimesUTC,
                 }
             ).then((response) => {
                 if (response.errors.length) {
@@ -162,76 +142,52 @@ export function VillageHQ({
                     return;
                 }
                 setChallengeDataState(response.data.challengeData);
-                openModal({
-                    header: 'Confirmation',
-                    text: response.data.response_message,
-                    ContentComponent: null,
-                    onConfirm: null,
-                });
+                setModalHeader("Confirmation");
+                setModalText(response.data.response_message);
+                setModalState("response_message");
             });
         }
     }
     const CancelChallenge = () => {
-        apiFetch(
-            villageAPI,
-            {
-                request: 'CancelChallenge',
-            }
-        ).then((response) => {
-            if (response.errors.length) {
-                handleErrors(response.errors);
-                return;
-            }
-            setChallengeDataState(response.data.challengeData);
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
+        if (modalState == "confirm_cancel_challenge") {
+            apiFetch(
+                villageAPI,
+                {
+                    request: 'CancelChallenge',
+                }
+            ).then((response) => {
+                if (response.errors.length) {
+                    handleErrors(response.errors);
+                    return;
+                }
+                setChallengeDataState(response.data.challengeData);
+                setModalHeader("Confirmation");
+                setModalText(response.data.response_message);
+                setModalState("response_message");
             });
-        });
-    }
-    const [challengeAcceptedFlag, setChallengeAcceptedFlag] = React.useState(false);
-    React.useEffect(() => {
-        if (challengeAcceptedFlag) {
-            ConfirmAcceptChallenge();
         }
-    }, [challengeAcceptedFlag]);
+        else {
+            setModalHeader("Confirmation");
+            setModalState("confirm_cancel_challenge");
+            setModalText("Are you sure you wish to cancel your pending challenge request?");
+        }
+    }
     const AcceptChallenge = (target_challenge) => {
-        challengeTarget.current = target_challenge;
-        openModal({
-            header: 'Accept Challenge',
-            text: "Select a time slot below to accept the challenge.",
-            ContentComponent: TimeGridResponse,
-            componentProps: ({
-                availableTimesUTC: JSON.parse(challengeTarget.current.selected_times),
-                selectedTimeUTC: selectedTimeUTC,
-                startHourUTC: luxon.DateTime.utc().minute === 0 ? luxon.DateTime.utc().hour + 12 : (luxon.DateTime.utc().hour + 13) % 24
-            }),
-            onConfirm: () => setChallengeAcceptedFlag(true),
-        });
+        setChallengeTarget(target_challenge);
+        setModalState("accept_challenge");
+        setModalHeader("Accept Challenge");
+        setModalText("Select a time slot below to accept the challenge.");
     }
     const ConfirmAcceptChallenge = () => {
-        setChallengeAcceptedFlag(false);
         if (!selectedTimeUTC) {
-            openModal({
-                header: 'Accept Challenge',
-                text: "Please verify that you have selected a time slot for the challenge.",
-                ContentComponent: TimeGridResponse,
-                componentProps: ({
-                    availableTimesUTC: JSON.parse(challengeTarget.current.selected_times),
-                    selectedTimeUTC: selectedTimeUTC,
-                    startHourUTC: luxon.DateTime.utc().minute === 0 ? luxon.DateTime.utc().hour + 12 : (luxon.DateTime.utc().hour + 13) % 24
-                }),
-                onConfirm: () => setChallengeAcceptedFlag(true),
-            });
+            setModalText("Select a slot to accept the challenge.");
         } else {
             apiFetch(
                 villageAPI,
                 {
                     request: 'AcceptChallenge',
-                    challenge_id: challengeTarget.current.request_id,
-                    time: selectedTimeUTC.current,
+                    challenge_id: challengeTarget.request_id,
+                    time: selectedTimeUTC,
                 }
             ).then((response) => {
                 if (response.errors.length) {
@@ -239,48 +195,111 @@ export function VillageHQ({
                     return;
                 }
                 setChallengeDataState(response.data.challengeData);
-                openModal({
-                    header: 'Confirmation',
-                    text: response.data.response_message,
-                    ContentComponent: null,
-                    onConfirm: null,
-                });
+                setModalHeader("Confirmation");
+                setModalText(response.data.response_message);
+                setModalState("response_message");
             });
         }
     }
     const LockChallenge = (target_challenge) => {
-        apiFetch(
-            villageAPI,
-            {
-                request: 'LockChallenge',
-                challenge_id: target_challenge.request_id,
-            }
-        ).then((response) => {
-            if (response.errors.length) {
-                handleErrors(response.errors);
-                return;
-            }
-            setChallengeDataState(response.data.challengeData);
-            openModal({
-                header: 'Confirmation',
-                text: response.data.response_message,
-                ContentComponent: null,
-                onConfirm: null,
+        if (modalState == "confirm_lock_challenge") {
+            apiFetch(
+                villageAPI,
+                {
+                    request: 'LockChallenge',
+                    challenge_id: challengeTarget.request_id,
+                }
+            ).then((response) => {
+                if (response.errors.length) {
+                    handleErrors(response.errors);
+                    return;
+                }
+                setChallengeDataState(response.data.challengeData);
+                setModalHeader("Confirmation");
+                setModalText(response.data.response_message);
+                setModalState("response_message");
             });
-        });
+        }
+        else {
+            setChallengeTarget(target_challenge);
+            setModalHeader("Confirmation");
+            setModalState("confirm_lock_challenge");
+            setModalText("Are you sure you want to lock in?\nYour actions will be restricted until the battle begins.");
+        }
+    }
+    const CancelChallengeSchedule = () => {
+        setSelectedTimesUTC([]);
     }
     const totalPopulation = populationData.reduce((acc, rank) => acc + rank.count, 0);
     const kage = seatDataState.find(seat => seat.seat_type === 'kage');
-
     return (
         <>
+            {modalState !== "closed" &&
+                <>
+                    <div className="modal_backdrop"></div>
+                    <div className="modal">
+                        <div className="modal_header">{modalHeader}</div>
+                        <div className="modal_text">{modalText}</div>
+                        {modalState == "confirm_resign" &&
+                            <>
+                                <div className="modal_confirm_button" onClick={() => Resign()}>confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "submit_challenge" &&
+                            <>
+                                <div className="schedule_challenge_subtext_wrapper">
+                                    <span className="schedule_challenge_subtext">Time slots are displayed in your local time.</span>
+                                    <span className="schedule_challenge_subtext">The seat holder will have 24 hours to choose one of your selected times.</span>
+                                    <span className="schedule_challenge_subtext">Your battle will be scheduled a minimum of 12 hours from the time of their selection.</span>
+                                </div>
+                                <TimeGrid
+                                    setSelectedTimesUTC={setSelectedTimesUTC}
+                                    startHourUTC={luxon.DateTime.fromObject({ hour: 0, zone: luxon.Settings.defaultZoneName }).toUTC().hour}
+                                />
+                                <div className="modal_confirm_button" onClick={() => ConfirmSubmitChallenge()}>confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "accept_challenge" &&
+                            <>
+                                <div className="schedule_challenge_subtext_wrapper">
+                                    <span className="schedule_challenge_subtext">Time slots are displayed in your local time.</span>
+                                    <span className="schedule_challenge_subtext">The first slot below is set a minimum 12 hours from the current time.</span>
+                                </div>
+                                <TimeGridResponse
+                                    availableTimesUTC={JSON.parse(challengeTarget.selected_times)}
+                                    setSelectedTimeUTC={setSelectedTimeUTC}
+                                    startHourUTC={(luxon.DateTime.utc().minute === 0 ? luxon.DateTime.utc().hour + 12 : luxon.DateTime.utc().hour + 13) % 24}
+                                />
+                                <div className="modal_confirm_button" onClick={() => ConfirmAcceptChallenge()}>confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "confirm_cancel_challenge" &&
+                            <>
+                                <div className="modal_confirm_button" onClick={() => CancelChallenge()}>confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "confirm_lock_challenge" &&
+                            <>
+                                <div className="modal_confirm_button" onClick={() => LockChallenge()}>confirm</div>
+                                <div className="modal_cancel_button" onClick={() => setModalState("closed")}>cancel</div>
+                            </>
+                        }
+                        {modalState == "response_message" &&
+                            <div className="modal_close_button" onClick={() => setModalState("closed")}>close</div>
+                        }
+                    </div>
+                </>
+            }
             <ChallengeContainer
                 playerID={playerID}
                 challengeDataState={challengeDataState}
                 CancelChallenge={CancelChallenge}
                 AcceptChallenge={AcceptChallenge}
                 LockChallenge={LockChallenge}
-                openModal={openModal}
             />
             <div className="hq_container">
                 <div className="row first">
@@ -340,17 +359,7 @@ export function VillageHQ({
                                     {kage.is_provisional ? kage.seat_title + ": " + kage.provisional_days_label : kage.seat_title + " of " + villageName}
                                 </div>
                                 {kage.seat_id && kage.seat_id == playerSeatState.seat_id &&
-                                    <div className="kage_resign_button"
-                                        onClick={
-                                            () => openModal({
-                                                header: 'Confirmation',
-                                                text: 'Are you sure you wish to resign from your current position?',
-                                                ContentComponent: null,
-                                                onConfirm: () => Resign(),
-                                            })
-                                        }>
-                                        resign
-                                    </div>
+                                    <div className="kage_resign_button" onClick={() => Resign()}>resign</div>
                                 }
                                 {!kage.seat_id &&
                                     <div className="kage_claim_button" onClick={() => ClaimSeat("kage")}>claim</div>
@@ -375,12 +384,7 @@ export function VillageHQ({
                                             </div>
                                             <div className="elder_name">{elder.user_name ? <a href={"/?id=6&user=" + elder.user_name}>{elder.user_name}</a> : "---"}</div>
                                             {(elder.seat_id && elder.seat_id == playerSeatState.seat_id) &&
-                                                <div className="elder_resign_button" onClick={() => openModal({
-                                                    header: 'Confirmation',
-                                                    text: 'Are you sure you wish to resign from your current position?',
-                                                    ContentComponent: null,
-                                                    onConfirm: () => Resign(),
-                                                })}>resign</div>
+                                                <div className="elder_resign_button" onClick={() => Resign()}>resign</div>
                                             }
                                             {!elder.seat_id &&
                                                 <div className={playerSeatState.seat_id ? "elder_claim_button disabled" : "elder_claim_button"} onClick={playerSeatState.seat_id ? null : () => ClaimSeat("elder")}>claim</div>
@@ -436,6 +440,15 @@ export function VillageHQ({
                                 </div>
                             </div>
                             <div className="village_policy_penalty_container">
+                                {policyDisplay.resources.map((resource, index) => (
+                                    <div key={index} className="policy_resource_item">
+                                        <svg width="16" height="16" viewBox="0 0 100 100">
+                                            <polygon points="25,20 50,45 25,70 0,45" fill="#414b8c" />
+                                            <polygon points="25,0 50,25 25,50 0,25" fill="#5964a6" />
+                                        </svg>
+                                        <div className="policy_resource_text">{resource}</div>
+                                    </div>
+                                ))}
                                 {policyDisplay.penalties.map((penalty, index) => (
                                     <div key={index} className="policy_penalty_item">
                                         <svg width="16" height="16" viewBox="0 0 100 100">
@@ -536,7 +549,7 @@ export function VillageHQ({
     );
 }
 
-export const ChallengeContainer = ({ playerID, challengeDataState, CancelChallenge, AcceptChallenge, LockChallenge, openModal }) => {
+export const ChallengeContainer = ({ playerID, challengeDataState, CancelChallenge, AcceptChallenge, LockChallenge }) => {
     return (
         <>
             {challengeDataState.length > 0 &&
@@ -565,20 +578,10 @@ export const ChallengeContainer = ({ playerID, challengeDataState, CancelChallen
                                             <div className="challenge_button lock disabled">lock in<img src="/images/v2/icons/unlocked.png" /></div>
                                         }
                                         {(challenge.start_time && luxon.DateTime.fromSeconds(challenge.start_time).toLocal() <= luxon.DateTime.local() && !challenge.challenger_locked) &&
-                                            <div className="challenge_button lock" onClick={() => openModal({
-                                                header: 'Confirmation',
-                                                text: "Are you sure you want to lock in?\nYour actions will be restricted until the battle begins.",
-                                                ContentComponent: null,
-                                                onConfirm: () => LockChallenge(challenge),
-                                            })}>lock in<img src="/images/v2/icons/unlocked.png" /></div>
+                                            <div className="challenge_button lock" onClick={() => LockChallenge(challenge)}>lock in<img src="/images/v2/icons/unlocked.png" /></div>
                                         }
                                         {(challenge.start_time == null) &&
-                                            <div className="challenge_button cancel" onClick={() => openModal({
-                                                header: 'Confirmation',
-                                                text: "Are you sure you wish to cancel your pending challenge request?",
-                                                ContentComponent: null,
-                                                onConfirm: () => CancelChallenge(),
-                                            })}>cancel</div>
+                                            <div className="challenge_button cancel" onClick={() => CancelChallenge()}>cancel</div>
                                         }
                                         {(challenge.challenger_locked) &&
                                             <div className="challenge_button locked">locked in<img src="/images/v2/icons/locked.png" /></div>
@@ -608,12 +611,7 @@ export const ChallengeContainer = ({ playerID, challengeDataState, CancelChallen
                                             <div className="challenge_button lock disabled">lock in<img src="/images/v2/icons/unlocked.png" /></div>
                                         }
                                         {(challenge.start_time && luxon.DateTime.fromSeconds(challenge.start_time).toLocal() <= luxon.DateTime.local() && !challenge.seat_holder_locked) &&
-                                            <div className="challenge_button lock" onClick={() => openModal({
-                                                header: 'Confirmation',
-                                                text: "Are you sure you want to lock in?\nYour actions will be restricted until the battle begins.",
-                                                ContentComponent: null,
-                                                onConfirm: () => LockChallenge(challenge),
-                                            })}>lock in<img src="/images/v2/icons/unlocked.png" /></div>
+                                            <div className="challenge_button lock" onClick={() => LockChallenge(challenge)}>lock in<img src="/images/v2/icons/unlocked.png" /></div>
                                         }
                                         {(challenge.start_time == null) &&
                                             <div className="challenge_button schedule" onClick={() => AcceptChallenge(challenge)}>schedule</div>
@@ -632,7 +630,7 @@ export const ChallengeContainer = ({ playerID, challengeDataState, CancelChallen
     );
 }
 
-export const TimeGrid = ({ selectedTimesUTC, startHourUTC = 0 }) => {
+export const TimeGrid = ({ setSelectedTimesUTC, startHourUTC = 0 }) => {
     const [selectedTimes, setSelectedTimes] = React.useState([]);
     const timeSlots = generateSlotsForTimeZone(startHourUTC);
 
@@ -654,7 +652,7 @@ export const TimeGrid = ({ selectedTimesUTC, startHourUTC = 0 }) => {
         }
 
         setSelectedTimes(newSelectedTimes);
-        selectedTimesUTC.current = convertSlotsToUTC(newSelectedTimes);
+        setSelectedTimesUTC(convertSlotsToUTC(newSelectedTimes));
     };
 
     function convertSlotsToUTC(times) {
@@ -667,11 +665,6 @@ export const TimeGrid = ({ selectedTimesUTC, startHourUTC = 0 }) => {
 
     return (
         <>
-            <div className="schedule_challenge_subtext_wrapper">
-                <span className="schedule_challenge_subtext">Time slots are displayed in your local time.</span>
-                <span className="schedule_challenge_subtext">The seat holder will have 24 hours to choose one of your selected times.</span>
-                <span className="schedule_challenge_subtext">Your battle will be scheduled a minimum of 12 hours from the time of their selection.</span>
-            </div>
             <div className="timeslot_container">
                 {timeSlots.map(time => (
                     <div key={time.toFormat('HH:mm')} onClick={() => toggleSlot(time)} className={selectedTimes.includes(time.toFormat('HH:mm')) ? "timeslot selected" : "timeslot"}>
@@ -692,7 +685,7 @@ export const TimeGrid = ({ selectedTimesUTC, startHourUTC = 0 }) => {
     );
 }
 
-export const TimeGridResponse = ({ availableTimesUTC, selectedTimeUTC, startHourUTC = 0 }) => {
+export const TimeGridResponse = ({ availableTimesUTC, setSelectedTimeUTC, startHourUTC = 0 }) => {
     const [selectedTime, setSelectedTime] = React.useState(null);
     const timeSlots = generateSlotsForTimeZone(startHourUTC);
 
@@ -708,19 +701,15 @@ export const TimeGridResponse = ({ availableTimesUTC, selectedTimeUTC, startHour
 
         if (selectedTime === formattedTimeLocal) {
             setSelectedTime(null);
-            selectedTimeUTC.current = null;
+            setSelectedTimeUTC(null);
         } else if (availableTimesUTC.includes(formattedTimeUTC)) {
             setSelectedTime(formattedTimeLocal);
-            selectedTimeUTC.current = formattedTimeUTC;
+            setSelectedTimeUTC(formattedTimeUTC);
         }
     };
 
     return (
         <>
-            <div className="schedule_challenge_subtext_wrapper">
-                <span className="schedule_challenge_subtext">Time slots are displayed in your local time.</span>
-                <span className="schedule_challenge_subtext">The first slot below is set a minimum 12 hours from the current time.</span>
-            </div>
             <div className="timeslot_container">
                 {timeSlots.map(time => {
                     const formattedTimeLocal = time.toFormat('HH:mm');
